feat: add firefox extension
