+++ extension/.gitignore
@@ -5,3 +5,4 @@ node_modules/
 packages/*
 platforms/chrome/build/*
 platforms/front/build/*
+platforms/firefox/build/*
