+++ extension/run/build.ts
@@ -3,11 +3,13 @@ import { isValidPlatform } from "@extension/shared/services/platform";
 import webpack from "webpack";
 
 import { getConfig as getChromeConfig } from "../platforms/chrome/webpack.config";
+import { getConfig as getFirefoxConfig } from "../platforms/firefox/webpack.config";
 import { getConfig as getFrontConfig } from "../platforms/front/webpack.config";
 
 const configPerPlatform: Record<PlatformType, any> = {
   chrome: getChromeConfig,
   front: getFrontConfig,
+  firefox: getFirefoxConfig,
 };
 
 async function main() {
