+++ extension/run/watch.ts
@@ -4,11 +4,13 @@ import webpack from "webpack";
 import WebpackDevServer from "webpack-dev-server";
 
 import { getConfig as getChromeConfig } from "../platforms/chrome/webpack.config";
+import { getConfig as getFirefoxConfig } from "../platforms/firefox/webpack.config";
 import { getConfig as getFrontConfig } from "../platforms/front/webpack.config";
 
 const configPerPlatform: Record<PlatformType, any> = {
   chrome: getChromeConfig,
   front: getFrontConfig,
+  firefox: getFirefoxConfig,
 };
 
 async function main() {
