+++ extension/shared/pageUtils.ts
