+++ extension/shared/tools/attachPageTextTool.ts
