+++ extension/shared/tools/tabActionTools.ts
@@ -1,8 +1,8 @@
 import { MCPError } from "@app/lib/actions/mcp_errors";
 import type { ToolHandlerResult } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { Err, Ok } from "@app/types/shared/result";
-import { sendTabActionMessage } from "@extension/platforms/chrome/messages";
 import { normalizeError } from "@extension/shared/lib/utils";
+import { sendTabActionMessage } from "@extension/shared/messages";
 
 // Registers tab manipulation tools with the MCP server.
