+++ extension/shared/tools/takeScreenshotOrAttachFileTool.ts
